feat: Post 합치기 및 api 공통폼에 맞게 수정
